src/seq2seq_sar/requirements.txt
torch
transformers
datasets
sentencepiece
pandas
numpy
scikit-learn
sacrebleu
rouge-score
pytest
pyyaml
matplotlib
seaborn
# Optional (recommended for faster GPU/distributed training)
# accelerate
# bitsandbytes (for 8-bit optimizers, if desired)
